requirements.txt
pyserial
numpy
seaborn
matplotlib
